started restructure: move globals, logging and database to dedicated files
